+++ packages/core/entrypoints.script.ts
@@ -1,0 +1,4 @@
+import { generateEntrypoints } from '../../scripts/entrypoints';
+import { input } from './rollup.config.js';
+
+generateEntrypoints(input);
