+++ packages/core/package.json
@@ -1,0 +1,101 @@
+{
+  "name": "@trpc/core",
+  "version": "10.44.1",
+  "description": "Internals of tRPC. Do not use directly.",
+  "author": "KATT",
+  "license": "MIT",
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "typings": "dist/index.d.ts",
+  "homepage": "https://trpc.io",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/trpc/trpc.git",
+    "directory": "packages/core"
+  },
+  "eslintConfig": {
+    "overrides": [
+      {
+        "files": [
+          "src"
+        ],
+        "rules": {
+          "explicit-module-boundary-types": "off",
+          "no-restricted-imports": [
+            "error",
+            "@trpc/core"
+          ]
+        }
+      }
+    ]
+  },
+  "scripts": {
+    "build": "rollup --config rollup.config.ts --configPlugin rollup-plugin-swc3",
+    "dev": "pnpm build --watch",
+    "codegen-entrypoints": "tsx entrypoints.script.ts",
+    "benchmark": "tsc --project tsconfig.benchmark.json",
+    "lint": "eslint --cache --ext \".js,.ts,.tsx\" --ignore-path ../../.gitignore --report-unused-disable-directives src",
+    "ts-watch": "tsc --watch"
+  },
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.js",
+      "default": "./dist/index.js"
+    },
+    "./observable": {
+      "import": "./dist/observable/index.mjs",
+      "require": "./dist/observable/index.js",
+      "default": "./dist/observable/index.js"
+    }
+  },
+  "files": [
+    "dist",
+    "src",
+    "README.md",
+    "package.json",
+    "observable",
+    "!**/*.test.*"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "devDependencies": {
+    "@fastify/websocket": "^7.1.2",
+    "@tanstack/react-query": "^5.0.0",
+    "@types/aws-lambda": "^8.10.97",
+    "@types/express": "^4.17.17",
+    "@types/hash-sum": "^1.0.0",
+    "@types/node": "^20.10.0",
+    "@types/react": "^18.2.33",
+    "@types/react-dom": "^18.2.14",
+    "@types/ws": "^8.2.0",
+    "@web3-storage/multipart-parser": "^1.0.0",
+    "aws-lambda": "^1.0.7",
+    "devalue": "^4.0.0",
+    "eslint": "^8.40.0",
+    "express": "^4.17.1",
+    "fastify": "^4.13.0",
+    "fastify-plugin": "^4.5.0",
+    "hash-sum": "^2.0.0",
+    "myzod": "^1.3.1",
+    "next": "^14.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "rollup": "^2.79.1",
+    "superjson": "^1.12.4",
+    "superstruct": "^1.0.0",
+    "tslib": "^2.5.0",
+    "tsx": "^4.0.0",
+    "typescript": "^5.3.3",
+    "valibot": "^0.24.0",
+    "vitest": "^0.32.0",
+    "ws": "^8.0.0",
+    "yup": "^1.0.0",
+    "zod": "^3.0.0"
+  },
+  "funding": [
+    "https://trpc.io/sponsor"
+  ]
+}
