+++ packages/core/src/observable/index.ts
@@ -1,0 +1,4 @@
+export * from './observable';
+export * from './types';
+export * from './operators';
+export { observableToPromise } from './internals/observableToPromise';
