+++ packages/core/src/observable/internals/identity.ts
