+++ packages/core/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src"],
+  "compilerOptions": {
+    // https://github.com/tsconfig/bases/blob/main/bases/node18.json
+    "lib": ["es2023", "DOM", "DOM.Iterable"],
+    "target": "es2022",
+    "module": "Node16",
+    "moduleResolution": "Node16"
+  }
+}
