+++ packages/server/src/observable/index.ts
@@ -1,4 +1,1 @@
-export * from './observable';
-export * from './types';
-export * from './operators';
-export { observableToPromise } from './internals/observableToPromise';
+export * from '@trpc/core/observable';
